Paso 41(Update 5.7-5.8)

+++ proyecto13/database/migrations/2022_11_27_202812_create_user_profiles_table.php
@@ -14,16 +14,16 @@ class CreateUserProfilesTable extends Migration
     public function up()
     {
         Schema::create('user_profiles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('bio', 1000);
             $table->string('twitter')->nullable();
-            $table->unsignedInteger('profession_id')
+            $table->unsignedBigInteger('profession_id')
                 ->nullable();
             $table->foreign('profession_id')
                 ->references('id')
                 ->on('professions');
-            $table->unsignedInteger('user_id') ->unique() ;		//no puede existir dos perfiles asociados al mismo usuario
+            $table->unsignedBigInteger('user_id') ->unique() ;		//no puede existir dos perfiles asociados al mismo usuario
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
