+++ proyecto13/database/migrations/2022_12_14_175205_add_team_id_to_users.php
@@ -15,7 +15,7 @@
 		public function up()
 		{
 			Schema::table('users', function (Blueprint $table) {
-				$table->unsignedInteger('team_id')
+				$table->unsignedBigInteger('team_id')
 					->nullable()->after('id');
 				$table->foreign('team_id')
 					->references('id')
